Make the library locale-independent and fix enum UB at the C boundary
Two defects that only matter now that FlatLand is a library rather than a
program, both found while auditing for Windows.

**Locale.** Every format FlatLand reads writes numbers with a '.', but strtod
uses the decimal separator of the process's LC_NUMERIC locale. The CLI has always
guarded this with setlocale(LC_ALL,"C") in main(); the library had no equivalent,
so loading a mesh from a host process in a comma-decimal locale failed outright:

    setlocale(LC_ALL, "de_DE.utf8");
    fl_mesh_load("cube.obj", &m);
    -> mesh 'cube.obj' line 2: vertex needs 3 coordinates, found 1

That is not a hypothetical. MATLAB and many desktop Python applications call
setlocale(LC_ALL,"") at startup, and comma-decimal locales are the default across
most of Europe — so this made the bindings unusable for a large fraction of their
intended users. It failed loudly rather than silently only because the strict
row parsing added earlier rejects a short vertex line; before that it would have
dropped rows and returned a plausible wrong answer.

A shared library must not call setlocale itself: that is global process state it
does not own, and changing it would affect every other thread. CNumericScope
instead switches only the calling thread to the C locale for the duration of a
parse, via POSIX uselocale() or Windows _configthreadlocale(), and restores it
on the way out. A test asserts both halves — that a mesh parses correctly under
de_DE.utf8, and that the host's own locale is unchanged afterwards. It skips
cleanly where no comma-decimal locale is installed.

**Enum UB.** check_options() read options.precision and options.field_mode as
their enum types in order to reject invalid values, but reading an enum whose
stored value is outside the enumeration is undefined behaviour — and validating
exactly that is the function's entire job, since a foreign caller can put any bit
pattern in the struct. UBSan flagged it against the test that deliberately
passes 42. An optimizer entitled to assume fl_precision only ever holds 0 or 1
could have deleted the check. The raw bytes are now copied out as int32 and
compared as integers, at the validation site and at every dispatch on precision.

Also from the same audit: covered_pixels widens to long long, which is 32-bit
`long` on Windows and is reported as int64_t across the ABI.

+++ src/fl_project.hpp
@@ -22,7 +22,10 @@ struct ViewResult {
     T min_val = 0;
     T max_val = 0;
     double time_seconds = 0.0;
-    long covered_pixels = 0;
+    // long long, not long: `long` is 32-bit on Windows (LLP64), and the
+    // raster ceiling is 1e9 pixels — inside int32 today, but with no headroom,
+    // and the C ABI reports this field as int64_t.
+    long long covered_pixels = 0;
     int image_width = 0;
     int image_height = 0;
     std::string output_image;
